Modularización de las funciones fetch y la función addToTablen los componetes BenefitModal y CrudBenefit, ahora se encuentran en componentes separado
